PKG -- [six] Update implementation
