+++ packages/six-create-account/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@onflow/six",
+  "name": "@onflow/six-create-account",
   "version": "0.0.0",
-  "description": "Flow SDK Stored Interactions",
+  "description": "Flow SDK Stored Interaction - Create Account",
   "license": "Apache-2.0",
   "author": "Dapper Labs <dev@dapperlabs.com>",
   "homepage": "https://onflow.org",
@@ -22,22 +22,21 @@
     "jest-esm-transformer": "1.0.0",
     "microbundle": "0.12.0-next.8"
   },
-  "source": "src/sdk.js",
-  "main": "dist/sdk.js",
-  "module": "dist/sdk.module.js",
-  "unpkg": "dist/sdk.umd.js",
+  "type": "module",
+  "source": "src/six.js",
+  "main": "dist/six.js",
+  "module": "dist/six.module.js",
+  "unpkg": "dist/six.umd.js",
   "scripts": {
     "prepublishOnly": "npm test && npm run build",
     "test": "jest",
     "build": "microbundle",
     "test:watch": "jest --watch",
-    "start": "microbundle watch"
+    "start": "microbundle watch",
+    "hash": "node src/hash-six-create-account.js"
   },
   "dependencies": {
-    "@onflow/decode": "0.0.6",
-    "@onflow/encode": "0.0.4",
-    "@onflow/interaction": "0.0.8",
-    "@onflow/send": "0.0.13",
-    "@qvvg/templar": "0.0.1"
+    "@onflow/sdk": "0.0.23",
+    "@onflow/types": "0.0.2"
   }
 }
